restaurant controller security

+++ app/Http/Controllers/Admin/RestaurantController.php
@@ -87,8 +87,6 @@ class RestaurantController extends Controller
         }
 
         return redirect()->route('admin.home');
-
-
     }
 
     /**
@@ -100,9 +98,13 @@ class RestaurantController extends Controller
     public function show(Restaurant $restaurant)
 
     {
-        $my_restaurant = $restaurant;
-        $user_auth = Auth::user();
-        return view('admin.restaurant.show', compact('my_restaurant', 'user_auth'));
+        if(Auth::user()->id == $restaurant->user->id){
+            $my_restaurant = $restaurant;
+            $user_auth = Auth::user();
+            return view('admin.restaurant.show', compact('my_restaurant', 'user_auth'));
+        }else{
+            abort(404);
+        }
     }
 
 
@@ -114,9 +116,12 @@ class RestaurantController extends Controller
      */
     public function edit(Restaurant $restaurant)
     {
-        $tipologies = Tipology::all();
-
-        return view('admin.restaurant.edit', compact('restaurant', 'tipologies'));
+        if(Auth::user()->id == $restaurant->user->id){
+            $tipologies = Tipology::all();
+            return view('admin.restaurant.edit', compact('restaurant', 'tipologies'));
+        }else{
+            abort(404);
+        }
     }
 
     /**
@@ -128,56 +133,61 @@ class RestaurantController extends Controller
      */
     public function update(Request $request, Restaurant $restaurant)
     {
-        /*
-            VALIDATION
-        */
-        $request->validate([
-            'name'=> 'required|min:2|max:50',
-            'address' => "required|min:2|max:255",
-            'phone_number' => [
-                'required',
-                'digits_between:7,15',
-                'numeric',
-                 Rule::unique('restaurants')->ignore($restaurant)
-            ],
-            'tipologies' => 'exists:tipologies,id|required_without_all',
-            'image' => 'nullable|mimes:jpg,jpeg,png,bmp,svg|max:5000',
-        ],[
-            'required'=> 'Questo campo è obbligatorio',
-            'max'=> 'Massimo :max caratteri concessi',
-            'min'=> 'Minimo :min caratteri richiesti',
-            'digits_between' => 'Inserisci un numero di telefono valido',
-            'numeric' => 'Inserisci solo numeri',
-            'exists' => 'valore non valido',
-            'required_without_all' => 'Seleziona almeno una casella',
-            'mimes' => 'I formati supportati sono: jpg,jpeg,png,bmp,svg',
-            'image.max' => 'Il file inserito eccede le misure massime consentite(5000kb)',
-            'phone_number.unique'=> 'il numero inserito è già esistente'
-        ]);
-
-
-        $data = $request->all();
-        $data['slug'] = Str::slug($data['name'], '-');
-
-        /* IMAGE */
-        if(array_key_exists('image', $data)){
-            if($restaurant->image){
-                Storage::delete($restaurant->image);
+        if(Auth::user()->id == $restaurant->user->id){
+            /*
+                VALIDATION
+            */
+            $request->validate([
+                'name'=> 'required|min:2|max:50',
+                'address' => "required|min:2|max:255",
+                'phone_number' => [
+                    'required',
+                    'digits_between:7,15',
+                    'numeric',
+                     Rule::unique('restaurants')->ignore($restaurant)
+                ],
+                'tipologies' => 'exists:tipologies,id|required_without_all',
+                'image' => 'nullable|mimes:jpg,jpeg,png,bmp,svg|max:5000',
+            ],[
+                'required'=> 'Questo campo è obbligatorio',
+                'max'=> 'Massimo :max caratteri concessi',
+                'min'=> 'Minimo :min caratteri richiesti',
+                'digits_between' => 'Inserisci un numero di telefono valido',
+                'numeric' => 'Inserisci solo numeri',
+                'exists' => 'valore non valido',
+                'required_without_all' => 'Seleziona almeno una casella',
+                'mimes' => 'I formati supportati sono: jpg,jpeg,png,bmp,svg',
+                'image.max' => 'Il file inserito eccede le misure massime consentite(5000kb)',
+                'phone_number.unique'=> 'il numero inserito è già esistente'
+            ]);
+    
+    
+            $data = $request->all();
+            $data['slug'] = Str::slug($data['name'], '-');
+    
+            /* IMAGE */
+            if(array_key_exists('image', $data)){
+                if($restaurant->image){
+                    Storage::delete($restaurant->image);
+                }
+                $image = Storage::put('restaurant-image', $data['image']);
+                $data['image'] = $image;
             }
-            $image = Storage::put('restaurant-image', $data['image']);
-            $data['image'] = $image;
-        }
-
-        /*UPDATING TIPOLOGIES TABLE TOO*/
-        if( array_key_exists( 'tipologies', $data ) ){
-            $restaurant->tipologies()->sync($data['tipologies']);
-        }else{
-            $restaurant->tipologies()->detach();
-        }
-
-        $restaurant->update( $data );
-
-        return redirect()->route('admin.home');
+    
+            /*UPDATING TIPOLOGIES TABLE TOO*/
+            if( array_key_exists( 'tipologies', $data ) ){
+                $restaurant->tipologies()->sync($data['tipologies']);
+            }else{
+                $restaurant->tipologies()->detach();
+            }
+    
+            $restaurant->update( $data );
+    
+            return redirect()->route('admin.home');
+        } 
+        else{
+            abort(404);
+        }       
     }
 
     /**
@@ -186,18 +196,23 @@ class RestaurantController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Restaurant $restaurant)
     {
-        $restaurant = Restaurant::find($id);
-
-        if($restaurant->image){
-            Storage::delete($restaurant->image);
+        if(!$restaurant){
+            abort(404);
+        }elseif(Auth::user()->id == $restaurant->user_id){
+            if($restaurant->image){
+                Storage::delete($restaurant->image);
+            }
+    
+            $restaurant->tipologies()->detach();
+            $restaurant->products()->delete();
+            $restaurant->delete();
+    
+            return redirect()->route('admin.home')->with('deleted', $restaurant->name);
+        }else{
+            abort(404);
         }
 
-        $restaurant->tipologies()->detach();
-        $restaurant->products()->delete();
-        $restaurant->delete();
-
-        return redirect()->route('admin.home')->with('deleted', $restaurant->name);
     }
 }
